requirements.txt
langchain==0.3.1
langchain-core==0.3.5
langchain-cohere==0.3.0
langgraph==0.2.22
pydantic==2.9.2
python-dotenv==1.0.1
requests==2.32.3
Flask==3.0.3

# PostgreSQL persistence dependencies
langgraph-checkpoint-postgres
psycopg[binary,pool]
